move take a photo to scanning a receipt
