+++ package.json
@@ -21,11 +21,12 @@
     "@react-navigation/material-top-tabs": "^6.6.14",
     "@react-navigation/native": "^6.0.2",
     "dotenv": "^16.4.5",
-    "expo": "~51.0.28",
+    "expo": "^51.0.38",
     "expo-constants": "~16.0.2",
     "expo-font": "~12.0.9",
+    "expo-image-picker": "~15.0.7",
     "expo-linking": "~6.3.1",
-    "expo-router": "~3.5.23",
+    "expo-router": "^3.5.23",
     "expo-splash-screen": "~0.27.5",
     "expo-status-bar": "~1.12.1",
     "expo-system-ui": "~3.0.7",
@@ -35,17 +36,16 @@
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "react-native": "0.74.5",
-    "react-native-document-scanner-plugin": "^1.0.1",
+    "react-native-document-scanner-plugin": "^0.1.2",
     "react-native-gesture-handler": "~2.16.1",
     "react-native-image-resizer": "^1.4.5",
+    "react-native-pager-view": "6.3.0",
     "react-native-reanimated": "~3.10.1",
     "react-native-safe-area-context": "4.10.5",
     "react-native-screens": "3.31.1",
     "react-native-tab-view": "^3.5.2",
     "react-native-web": "~0.19.10",
-    "tesseract.js": "^5.1.1",
-    "react-native-pager-view": "6.3.0",
-    "expo-image-picker": "~15.0.7"
+    "tesseract.js": "^5.1.1"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
